requirements-win7.txt
pyinstaller==5.13.2
pystray==0.19.5
keyboard==0.13.5
pynput==1.7.6
pillow==9.5.0
setuptools==69.5.1
